requirements.txt
lxml
validators
requests
pyopenssl
